requirements.txt
et-xmlfile==1.0.1
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
selenium==3.141.0
six==1.15.0
urllib3==1.26.5
xlrd==2.0.1
coleta-dadosjusbr==0.3
protobuf==3.18.0